Change to double the data Calculate distances

+++ graph01/src/orbitEngine/Body.java
@@ -6,34 +6,34 @@ package orbitEngine;
  */
 public class Body {
 
-    public static final float G_UNIVERSAL = (float) 6.6740831e-11; // m^3/(Kg*s^2)
+    public static final double G_UNIVERSAL = 6.6740831e-11; // m^3/(Kg*s^2)
 
     String name;
-    float mass;
-    float g_mass;
-    float radius;
+    double mass;
+    double g_mass;
+    double radius;
 
     // Position by axis r(t)
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 
     // Speed by axis: v(t)= d(r(t))/dt 
-    float vx;
-    float vy;
-    float vz;
+    double vx;
+    double vy;
+    double vz;
 
     // Gravity acceleration by axis: a(t)= d(v(t))/dt = d2(r(t))/dt
-    float ax[];
-    float ay[];
-    float az[];
+    double ax[];
+    double ay[];
+    double az[];
 
     // Jerk by axis: j(t)=d(a(t))/dt = d2(v(t))/dt = d3(r(t))/dt
-    float jx[];
-    float jy[];
-    float jz[];
+    double jx[];
+    double jy[];
+    double jz[];
 
-    public Body(String name, float mass, float radius, float x, float y, float z, float vx, float vy, float vz) {
+    public Body(String name, double mass, double radius, double x, double y, double z, double vx, double vy, double vz) {
         this.name = name;
         this.mass = mass;
         this.g_mass = mass * G_UNIVERSAL;
@@ -46,13 +46,13 @@ public class Body {
         this.vz = vz;
     }
 
-    public void addGravity(float[] ax, float[] ay, float[] az) {
+    public void addGravity(double[] ax, double[] ay, double[] az) {
         this.ax = ax;
         this.ay = ay;
         this.az = az;
     }
 
-    public void addJerk(float[] jx, float[] jy, float[] jz) {
+    public void addJerk(double[] jx, double[] jy, double[] jz) {
         this.jx = jx;
         this.jy = jy;
         this.jz = jz;
